normalizing beta, setting beta[0]=0

+++ segment.py
@@ -9,7 +9,7 @@ def parseCommandLine():
     parser.add_argument("-t", "--train", dest="train", default="", help="Train filename")
     parser.add_argument("-d", "--dev", dest="dev", default="", help="development filename")
     parser.add_argument("-b", "--beta", dest="beta", default=1.3, type=float, help="Higher beta makes longer sequences less likely")
-    parser.add_argument("-c", "--offset", dest="offset", default=0, type=int, help="Set c higher than 0 if short segments are unlikely")
+    parser.add_argument("-c", "--offset", dest="offset", default=0, type=float, help="Set c higher than 0 if short segments are unlikely")
     parser.add_argument("-S", "--maxseglength", dest="MAX_SEG_LENGTH", default=10, type=int, help="maximum segment length")
     parser.add_argument("-i", "--iterations", dest="ITERATIONS", default=6, type=int, help="EM iterations")
